[MachineOutliner][NFC] Remove buildCandidateList and replace with findCandidates
More refactoring.

Since the pruning logic has changed, and the candidate list is gone,
everything can be sunk into findCandidates.

We no longer need to keep track of the length of the longest substring, so we
can drop all of that logic as well.

After this, we just find all of the candidates and move to outlining.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@348428 91177308-0d34-0410-b5e6-96231b3b80d8