[X86] Correct the MaskVT for avx512 gather/scatter intrinsics to use the min of the number of index and data elements.
When the result type is v2i64/v2f64 and the index element size is i32, the index vector has two unused elements making the type v4i32. The mask VT should match the number of memory accesses that will be made.

This is consistent with the isel patterns used for the target independent gather/scatter intrinsic.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@350687 91177308-0d34-0410-b5e6-96231b3b80d8