[VFS] Add getter for mapping entries.
When generating a reproducer in LLDB we build up the mapping but don't
immediately copy over the files on the file system.

Rather than keeping a separate data structure with real and virtual
paths, we might as well reuse the entries already stored in the
YAMLVFSWriter to lazily copy over the files when needed.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@351266 91177308-0d34-0410-b5e6-96231b3b80d8